racso_calendar_export/requirements.txt
xlwt>=1.3.0
pyobjc-framework-EventKit>=10.0
